requirements.txt
numpy<2.0
pandas
polars
tqdm
lightgbm
mypy
cudf
faiss-cpu
scikit-learn
torch
torch_geometric